Feature parity with CC:T 1.85.2

+++ rom/programs/rename.lua
@@ -7,8 +7,12 @@ end
 local sSource = shell.resolve( tArgs[1] )
 local sDest = shell.resolve( tArgs[2] )
 
-if fs.exists( sDest ) then
+if not fs.exists( sSource ) then
+    printError( "No matching files" )
+elseif fs.exists( sDest ) then
     printError( "Destination exists" )
+elseif fs.isReadOnly( sDest ) then
+    printError( "Destination is read-only" )
+else
+    fs.move( sSource, sDest )
 end
-
-fs.move( sSource, sDest )
